Move helper functions in api module into the internal package, extract functions only used in other modules
