+++ detekt-cli/src/main/kotlin/io/gitlab/arturbosch/detekt/cli/console/ReportFormatting.kt
@@ -1,0 +1,9 @@
+package io.gitlab.arturbosch.detekt.cli.console
+
+/**
+ * @author Artur Bosch
+ */
+
+const val PREFIX = "\t- "
+
+fun Any.format(prefix: String = "", suffix: String = "\n") = "$prefix$this$suffix"
