+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/Junk.kt
@@ -5,6 +5,7 @@ import org.jetbrains.kotlin.psi.KtFile
 import java.io.PrintStream
 import java.nio.file.Files
 import java.nio.file.Path
+import java.util.HashMap
 import java.util.stream.Collectors
 import java.util.stream.Stream
 
@@ -30,4 +31,14 @@ fun MutableMap<String, List<Finding>>.mergeSmells(other: Map<String, List<Findin
 fun Throwable.printStacktraceRecursively(logger: PrintStream) {
 	stackTrace.forEach { logger.println(it) }
 	cause?.printStacktraceRecursively(logger)
+}
+
+fun <K, V> List<Pair<K, List<V>>>.toMergedMap(): Map<K, List<V>> {
+	val map = HashMap<K, MutableList<V>>()
+	for ((key, values) in this) {
+		map.merge(key, values.toMutableList()) { l1, l2 ->
+			l1.apply { addAll(l2) }
+		}
+	}
+	return map
 }
